fix(docs): resolve canonical Docs standards findings (#257)
